+++ packages/capillary-ui/src/Components/lineinputs/textbox.tsx
@@ -156,6 +156,11 @@ export class Textbox extends LabeledInputControl<TextboxProps> {
             cursor: not-allowed;
         }
 
+        & > input:required:invalid:not(:disabled):not([readonly]):not([aria-invalid="true"]) {
+            outline: 1px dashed var(--required-color);
+            outline-offset: 3px;
+        }
+
         & > input:focus-visible {
             outline: 2px solid transparent;
             outline-offset: 1px;